+++ qobuz-player-tui/src/app.rs
@@ -86,7 +86,7 @@ pub enum Output {
     Popup(Popup),
     PopPopupUpdateFavorites,
     AddTrackToPlaylistPopup(Track),
-    AddTrackToPlaylistAndPopPopup((u32, u32)), // TODO: Add a type
+    AddTrackToPlaylistAndPopPopup((u32, i64)), // TODO: Add a type
 }
 
 #[derive(Default, PartialEq)]
